refactor(fn): move fn functions to separate files
